requirements.txt
numpy==1.21.5
pandas==1.5.1
pillow==9.3.0
plotly==5.11.0
streamlit==1.15.0
tensorflow==2.9.1
keras==2.9.0